customer order test refactor: move general db functionality tests to a separate test file
