+++ apps/web-client/src/lib/db/cr-sqlite/__tests__/customer-orders.test.ts
@@ -2,7 +2,7 @@ import { describe, it, expect, beforeEach } from "vitest";
 
 import type { DB, Customer } from "../types";
 
-import { getDB, initializeDB } from "../db";
+import { markCustomerOrderAsReceived, getRandomDb } from "./lib";
 
 import {
 	getAllCustomers,
@@ -15,21 +15,9 @@ import {
 	isDisplayIdUnique,
 	markCustomerOrderLineAsCollected
 } from "../customers";
-// import { createSupplierOrder, getPossibleSupplierOrderLines } from "../suppliers";
-import { markCustomerOrderAsReceived, getRandomDb, getRandomDbs, syncDBs } from "./lib";
 import { associatePublisher, createSupplierOrder, getPlacedSupplierOrders, upsertSupplier } from "../suppliers";
 import { upsertBook } from "../books";
 import { addOrderLinesToReconciliationOrder, createReconciliationOrder, finalizeReconciliationOrder } from "../order-reconciliation";
-
-describe("Db creation tests", () => {
-	it("should allow initializing a database", async () => {
-		const randomTestRunId = Math.floor(Math.random() * 100000000);
-		const db = await getDB("init-db-test" + randomTestRunId);
-		await expect(getAllCustomers(db)).rejects.toThrow();
-		await initializeDB(db);
-		expect((await getAllCustomers(db)).length).toBe(0);
-	});
-});
 
 describe("Customer order tests", () => {
 	let db: DB;
@@ -165,45 +153,6 @@ describe("Customer order tests", () => {
 
 		const [line2] = await getCustomerOrderLines(db, 1);
 		expect(Date.now() - line2.collected.getTime()).toBeLessThan(300);
-	});
-});
-
-describe("Customer order tests", () => {
-	let db1: DB, db2: DB;
-
-	beforeEach(async () => ([db1, db2] = await getRandomDbs()));
-
-	it("Should sync customer creation", async () => {
-		// We create one customer in db1 and a different one in db2
-		let db1Customers: Customer[], db2Customers: Customer[];
-		await upsertCustomer(db1, { fullname: "John Doe", id: 1, email: "john@example.com", deposit: 13.2, displayId: "1" });
-		await upsertCustomer(db2, { fullname: "Jane Doe", id: 2, displayId: "2" });
-		[db1Customers, db2Customers] = await Promise.all([getAllCustomers(db1), getAllCustomers(db2)]);
-		expect(db1Customers.length).toBe(1);
-		expect(db2Customers.length).toBe(1);
-
-		// This tests for a regression we had: we want to ensure that the sync won't update the `updated_at` field
-		// - it should be the same as for the original entry
-		await new Promise((resolve) => setTimeout(resolve, 1000));
-
-		await syncDBs(db1, db2);
-		expect((await getAllCustomers(db2)).length).toBe(2);
-		await syncDBs(db2, db1);
-
-		expect((await getAllCustomers(db1)).length).toBe(2);
-		[db1Customers, db2Customers] = await Promise.all([getAllCustomers(db1), getAllCustomers(db2)]);
-		expect(db1Customers).toMatchObject(db2Customers);
-	});
-
-	it("Should keep both updates done at the same time on different dbs", async () => {
-		// We create one customer in db1 and a different one in db2
-		await upsertCustomer(db1, { fullname: "John Doe", id: 1, email: "john@example.com", deposit: 13.2, displayId: "1" });
-		await syncDBs(db1, db2);
-		await upsertCustomer(db2, { fullname: "Jane Doe", id: 1, email: "jane@example.com", displayId: "1" });
-		await syncDBs(db2, db1);
-		const [db1Customers, db2Customers] = await Promise.all([getAllCustomers(db1), getAllCustomers(db2)]);
-		expect(db1Customers).toMatchObject(db2Customers);
-		expect(db1Customers).toMatchObject([{ fullname: "Jane Doe", id: 1, email: "jane@example.com", deposit: 13.2 }]);
 	});
 });
 
@@ -421,22 +370,3 @@ describe("isDisplayIdUnique function", () => {
 		expect(await isDisplayIdUnique(db, { id: 1, displayId: "1" })).toEqual(true);
 	});
 });
-
-// TODO: update this when we have a handler to getPlacedOrderLines
-// describe("Customer order status", () => {
-// 	let db: DB;
-// 	beforeEach(async () => {
-// 		db = await getRandomDb();
-// 		await createCustomerOrders(db);
-// 	});
-// 	it("can update the timestamp of when a customer order is placed (to supplier)", async () => {
-// 		const newOrderLines = await getPossibleSupplierOrderLines(db, 1);
-
-// 		await createSupplierOrder(db, newOrderLines);
-
-// 	const isbns = [...newOrders[0].lines, ...newOrders[1].lines].map((line) => line.isbn);
-// 	await markCustomerOrderAsReceived(db, isbns);
-// 	const books = await getCustomerBooks(db, 1);
-// 	expect(books[1].received).toBeInstanceOf(Date);
-// 	});
-// });
